Fix ioctl() argument type

+++ src/serial.c
@@ -132,7 +132,7 @@ int serial_set_parity(port_handle_t fd, int enable, int odd_parity)
 
 int serial_set_dtr(port_handle_t fd, int level)
 {
-    int command;
+    unsigned long command;
     const int dtr = TIOCM_DTR;
     if (level)
     {
@@ -147,7 +147,7 @@ int serial_set_dtr(port_handle_t fd, int level)
 
 int serial_set_rts(port_handle_t fd, int level)
 {
-    int command;
+    unsigned long command;
     const int rts = TIOCM_RTS;
     if (level)
     {
@@ -162,7 +162,7 @@ int serial_set_rts(port_handle_t fd, int level)
 
 int serial_set_txd(port_handle_t fd, int level)
 {
-    int command;
+    unsigned long command;
     if (level)
     {
         command = TIOCCBRK;
